+++ src/Admin/Entities/Exception/ArticleAlreadyExistsException.php
@@ -1,0 +1,36 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the Tests package.
+ *
+ * (c) Dev-Int Création <info@developpement-interessant.com>.
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Admin\Entities\Exception;
+
+use Shared\Entities\Exception\ExceptionSerializableTrait;
+
+final class ArticleAlreadyExistsException extends \DomainException implements \JsonSerializable
+{
+    use ExceptionSerializableTrait;
+
+    public const MESSAGE = 'Article already exists.';
+
+    public function __construct(?\Throwable $previous = null)
+    {
+        parent::__construct(self::MESSAGE, 0, $previous);
+    }
+
+    /**
+     * @return iterable<string, array<int, string>|int|string>
+     */
+    public function jsonSerialize(): iterable
+    {
+        return $this->toJson();
+    }
+}
