+++ src/Admin/UseCases/Article/CreateArticle/CreateArticleResponse.php
@@ -1,0 +1,23 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the Tests package.
+ *
+ * (c) Dev-Int Création <info@developpement-interessant.com>.
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Admin\UseCases\Article\CreateArticle;
+
+use Admin\Entities\Article\Article;
+
+final readonly class CreateArticleResponse
+{
+    public function __construct(public Article $article)
+    {
+    }
+}
