+++ src/Admin/UseCases/Gateway/ArticleRepository.php
@@ -1,0 +1,23 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the Tests package.
+ *
+ * (c) Dev-Int Création <info@developpement-interessant.com>.
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Admin\UseCases\Gateway;
+
+use Admin\Entities\Article\Article;
+
+interface ArticleRepository
+{
+    public function isExists(string $name): bool;
+
+    public function save(Article $article): void;
+}
