--- src/Article/Adapters/ORM/Resources/doctrine/Article.orm.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<doctrine-mapping xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping"
-                  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-                  xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping
-                          https://www.doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
-    <entity name="Domain\Model\Article\Article"
-            table="article"
-            repository-class="Infrastructure\Persistence\DoctrineOrm\Repositories\DoctrineArticleRepository">
-        <id name="articleId" type="integer" column="id">
-            <generator strategy="AUTO"/>
-        </id>
-        <field name="name" type="string" length="255" nullable="false" unique="true" />
-        <field name="supplier" type="string" length="255" nullable="false" />
-        <field name="unitStorage" type="string" length="50" nullable="false"/>
-        <field name="packaging" type="array" nullable="false" />
-        <field name="price" type="decimal" scale="6" precision="3" nullable="false" />
-        <field name="tva" type="string" nullable="false" />
-        <field name="minStock" type="decimal" scale="6" precision="3" nullable="false" />
-        <field name="zoneStorages" type="array" length="50" nullable="false"/>
-        <field name="familyLog" type="string" nullable="false" />
-        <field name="quantity" type="decimal" scale="6" precision="3" nullable="false" />
-        <field name="slug" type="string" nullable="false" unique="true" />
-        <field name="active" type="boolean" nullable="false" />
-    </entity>
-</doctrine-mapping>
